+++ src/registryctl/api/base.go
@@ -17,7 +17,7 @@ package api
 import (
 	"encoding/json"
 	"github.com/goharbor/harbor/src/lib/errors"
-	server_error "github.com/goharbor/harbor/src/server/error"
+	lib_http "github.com/goharbor/harbor/src/lib/http"
 	"net/http"
 )
 
@@ -38,7 +38,7 @@ func HandleBadRequest(w http.ResponseWriter, err error) {
 
 // HandleError ...
 func HandleError(w http.ResponseWriter, err error) {
-	server_error.SendError(w, err)
+	lib_http.SendError(w, err)
 }
 
 // WriteJSON response status code will be written automatically if there is an error
