+++ src/server/middleware/middleware.go
@@ -15,10 +15,10 @@
 package middleware
 
 import (
+	lib_http "github.com/goharbor/harbor/src/lib/http"
 	"net/http"
 
 	"github.com/goharbor/harbor/src/lib"
-	serror "github.com/goharbor/harbor/src/server/error"
 )
 
 // Middleware receives a handler and returns another handler.
@@ -63,7 +63,7 @@ func New(fn func(http.ResponseWriter, *http.Request, http.Handler), skippers ...
 func BeforeRequest(hook func(*http.Request) error, skippers ...Skipper) func(http.Handler) http.Handler {
 	return New(func(w http.ResponseWriter, r *http.Request, next http.Handler) {
 		if err := hook(lib.NopCloseRequest(r)); err != nil {
-			serror.SendError(w, err)
+			lib_http.SendError(w, err)
 			return
 		}
 
@@ -85,7 +85,7 @@ func AfterResponse(hook func(http.ResponseWriter, *http.Request, int) error, ski
 
 		if err := hook(res, r, res.StatusCode()); err != nil {
 			res.Reset()
-			serror.SendError(res, err)
+			lib_http.SendError(res, err)
 		}
 	}, skippers...)
 }
