+++ src/server/version.go
@@ -16,9 +16,9 @@ package server
 
 import (
 	"encoding/json"
+	lib_http "github.com/goharbor/harbor/src/lib/http"
 	"net/http"
 
-	serror "github.com/goharbor/harbor/src/server/error"
 	"github.com/goharbor/harbor/src/server/v2.0/route"
 )
 
@@ -34,6 +34,6 @@ type APIVersion struct {
 // GetAPIVersion returns the current supported API version
 func GetAPIVersion(w http.ResponseWriter, r *http.Request) {
 	if err := json.NewEncoder(w).Encode(&APIVersion{Version: version}); err != nil {
-		serror.SendError(w, err)
+		lib_http.SendError(w, err)
 	}
 }
